fix(review): address 2nd-round findings — Windows file-lock cleanup, 32-bit truncation
- Windows compat: drop(blob_file) before remove_file calls
- 32-bit safety: compare file sizes as u64, not usize (prevents
  truncation on 32-bit targets for containers >4GB)

+++ src-tauri/src/crypto.rs
@@ -210,8 +210,8 @@ pub fn stream_verify_header(
     ) as usize;
 
     // Guard against OOM from malicious/corrupted metadata size
-    let file_len = file.metadata()?.len() as usize;
-    if meta_len > file_len || meta_len > 50 * 1024 * 1024 {
+    let file_len = file.metadata()?.len();
+    if (meta_len as u64) > file_len || meta_len > 50 * 1024 * 1024 {
         return Err(CryptoError::InvalidFormat("Metadata size is invalid or too large".into()));
     }
